+++ R/ObjFnBinomMaxLH2.R
@@ -62,8 +62,6 @@ ObjFnBinomMaxLH2 <-
 #' Can be used as the
 #' objective function for \code{\link{MAPAssignActivity}},
 #' \code{\link{MAPAssignActivity1}},
-#' \code{\link{SparseAssignActivity}},
-# \code{\link{SparseAssignActivity1}}, ##### function removed
 #' \code{\link{SignaturePresenceTest}}
 #' and \code{\link{SignaturePresenceTest1}}.
 #' (Internally used by by \code{\link[nloptr]{nloptr}}.)