+++ R/ObjFnMultinomMaxLH.R
@@ -3,8 +3,6 @@
 #' Can be used as the
 #' objective function for \code{\link{MAPAssignActivity}},
 #' \code{\link{MAPAssignActivity1}},
-#' \code{\link{SparseAssignActivity}},
-# \code{\link{SparseAssignActivity1}},  ##### Function removed
 #' \code{\link{SignaturePresenceTest}}
 #' and \code{\link{SignaturePresenceTest1}}.
 #' (Internally used by by \code{\link[nloptr]{nloptr}}.)