+++ R/PresenceAssignActivity.R
@@ -79,8 +79,6 @@ PresenceAssignActivity <-
            seed                      = NULL,
            drop.low.mut.samples      = TRUE,
            use.forward.search        = FALSE,
-           sig.pres.test.nbinom.size = NULL,
-           sig.pres.test.p.thresh    = 0.05,
            save.files                = TRUE) {
     
     retval <- 
@@ -99,8 +97,6 @@ PresenceAssignActivity <-
                         max.subsets               = .Machine$double.xmax,
                         drop.low.mut.samples      = drop.low.mut.samples,
                         use.sig.presence.test     = TRUE,
-                        sig.pres.test.nbinom.size = sig.pres.test.nbinom.size,
-                        sig.pres.test.p.thresh    = sig.pres.test.p.thresh,
                         save.files                = save.files)
     return(retval)
   }