+++ NEWS.md
@@ -1,3 +1,6 @@
+# 3.0.1
+* Renamed exported function `PresenceAssignActivity` to `PresenceAssignSigActivity`.
+
 # 3.0.0
 * Added new internal function `DefaultPThresh`.